refactor: 공통으로 사용하는 type, api @dnd-academy/core로 이동
