+++ apps/admin/src/app/page.tsx
@@ -1,5 +1,7 @@
 import { redirect } from 'next/navigation';
 
+import { api, CurrentApplicantCount } from '@dnd-academy/core';
+
 import { auth } from '@/auth';
 import SignOut from '@/components/auth/SignOut';
 
@@ -10,18 +12,18 @@ async function MainPage() {
     redirect('/login');
   }
 
-  // TODO - 제거 예정 (임시 호출 테스트)
-  const response = await fetch(`${process.env.NEXT_PUBLIC_VERCEL_BLOB_HOST}/current_applicant_count.json`, {
+  const currentApplicantCountData = await api<CurrentApplicantCount>({
+    url: '/current_applicant_count.json',
     method: 'GET',
   });
-
-  const currentApplicantCountData = await response.json();
 
   console.log('currentApplicantCountData', currentApplicantCountData);
 
   return (
     <>
       <h1>DND - AdminPage</h1>
+      <div>{currentApplicantCountData.designer}</div>
+      <div>{currentApplicantCountData.developer}</div>
       <SignOut />
     </>
   );
