+++ apps/web/src/app/page.tsx
@@ -1,10 +1,9 @@
+import { api, CurrentApplicantCount } from '@dnd-academy/core';
+
 import HomePage from '@/components/pages/HomePage';
 import { getEventStatus } from '@/lib/apis/event';
 import { DEFAULT_METADATA } from '@/lib/constants/metadata';
-import { CurrentApplicantCount } from '@/lib/types/count';
 import { checkNumber } from '@/utils';
-
-import api from './api';
 
 export const metadata = DEFAULT_METADATA;
