+++ apps/web/src/components/organisms/ProjectCards/index.tsx
@@ -1,9 +1,8 @@
 import Image from 'next/image';
 import Link from 'next/link';
 
+import type { Project } from '@dnd-academy/core';
 import { SkillTag } from '@dnd-academy/ui';
-
-import { Project } from '@/lib/types/project';
 
 import styles from './index.module.scss';
