+++ apps/web/src/components/pages/ProjectsPage/index.tsx
@@ -1,7 +1,8 @@
+import type { Project } from '@dnd-academy/core';
+
 import Tags from '@/components/molecules/Tags';
 import ProjectCards from '@/components/organisms/ProjectCards';
 import { getProjectCount } from '@/lib/apis/project';
-import { Project } from '@/lib/types/project';
 
 type Props = {
   projects: Project[];
