+++ apps/web/src/lib/apis/event.ts
@@ -1,5 +1,6 @@
+import { EventStatus } from '@dnd-academy/core';
+
 import { eventStatusData } from '../assets/data';
-import { EventStatus } from '../types/event';
 
 // eslint-disable-next-line import/prefer-default-export
 export const getEventStatus = () => eventStatusData as EventStatus;
