+++ packages/core/src/@types/job.ts
