+++ packages/core/src/index.ts
@@ -1,2 +1,3 @@
 export * from './api';
+export type * from './types';
 export * from './utils';
