+++ packages/core/vite.config.mts
@@ -1,3 +1,5 @@
+import { resolve } from 'path';
+
 import { defineConfig } from 'vite';
 import dts from 'vite-plugin-dts';
 
@@ -11,7 +13,9 @@ export default defineConfig({
   ],
   build: {
     lib: {
-      entry: './src/index.ts',
+      entry: {
+        index: resolve(__dirname, 'src/index.ts'),
+      },
       formats: ['es', 'cjs'],
       name: '@dnd-academy/core',
       fileName: (format, entryName) => `${entryName}.${format === 'es' ? 'mjs' : 'js'}`,
